LayoutTests/imported/w3c/web-platform-tests/css/css-transitions/parsing/transition-property-computed-expected.txt


PASS Property transition-property value 'left, top, right, bottom' 
FAIL Property transition-property value 'one, two, three' assert_equals: expected "one, two, three" but got "three"
